+++ src/compiler/scala/tools/nsc/reporters/Reporter.scala
@@ -13,8 +13,8 @@ import scala.reflect.internal.util._
  * This describes the internal interface for issuing information, warnings and errors.
  * The only abstract method in this class must be info0.
  *
- * TODO: Move external clients (sbt/ide/partest) to reflect.internal.Reporter
- *       This interface should be considered private to the compiler.
+ * TODO: Move external clients (sbt/ide/partest) to reflect.internal.Reporter,
+ *       and remove this class.
  */
 abstract class Reporter extends scala.reflect.internal.Reporter {
   /** Informational messages. If `!force`, they may be suppressed. */
